requirements.txt
# Core PyTorch dependencies
torch>=2.8.0
torchvision>=0.19.0
torchaudio>=2.4.0
numpy>=1.24.0,<2.0.0

# Configuration and serialization
pyyaml>=6.0.0

# GPU monitoring
pynvml>=11.0.0
nvidia-ml-py>=11.450.51

# Server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Client dependencies
requests>=2.31.0

# HuggingFace models and evaluation workloads
transformers>=4.20.0
Pillow>=9.0.0
scipy>=1.10.0

# System monitoring
psutil>=5.9.0

# Progress bars
tqdm>=4.64.0

# Development and testing dependencies
pytest>=7.3.1
pytest-asyncio>=1.2.0
networkx>=3.0
packaging>=23.0

# Data analysis and visualization (for evaluation scripts)
matplotlib>=3.5.0
pandas>=2.0.0
seaborn>=0.13.0

# Distributed computing (for benchmarks)
ray>=2.0.0
